wip: fix ooc groupby
