+++ polars/polars-lazy/polars-pipe/src/executors/sinks/groupby/generic.rs
@@ -7,11 +7,11 @@ use num_traits::NumCast;
 use polars_arrow::trusted_len::PushUnchecked;
 use polars_core::export::ahash::RandomState;
 use polars_core::frame::row::AnyValueBuffer;
+use polars_core::hashing::hash_to_partition;
 use polars_core::prelude::*;
 use polars_core::series::SeriesPhysIter;
 use polars_core::utils::_set_partition_size;
 use polars_core::{IdBuildHasher, POOL};
-use polars_utils::hash_to_partition;
 use polars_utils::slice::GetSaferUnchecked;
 use polars_utils::unwrap::UnwrapUncheckedRelease;
 use rayon::prelude::*;
