+++ polars/polars-utils/src/functions.rs
@@ -12,13 +12,6 @@ pub fn flatten<T: Clone, R: AsRef<[T]>>(bufs: &[R], len: Option<usize>) -> Vec<T
 }
 
 #[inline]
-pub fn hash_to_partition(h: u64, n_partitions: usize) -> usize {
-    debug_assert!(n_partitions.is_power_of_two());
-    // n % 2^i = n & (2^i - 1)
-    h as usize & n_partitions.wrapping_sub(1)
-}
-
-#[inline]
 pub fn get_hash<T: Hash, B: BuildHasher>(value: T, hb: &B) -> u64 {
     let mut hasher = hb.build_hasher();
     value.hash(&mut hasher);
